move changelogs to a separate file, bump play services + support lib versions, cleanup code, remove some lint stuff

+++ app/src/main/java/net/mbonnin/arcanetracker/DeckEditorView.java
@@ -195,7 +195,7 @@ public class DeckEditorView extends RelativeLayout {
         ViewManager.Params params = new ViewManager.Params();
         params.x = 0;
         params.y = 0;
-        params.w = (int) (viewManager.getUsableWidth());
+        params.w = viewManager.getUsableWidth();
         params.h = viewManager.getUsableHeight();
 
         deckEditorView.setDeck(deck);
